Add files via upload
Used Daily Delhi Climate Test data and used exponentially weighted moving average with alpha value 0.1 (beta= 0.9) to understand the time series data. If the beta value is high then we are giving more weightage to the previous data. For example, if beta=0.9 then average = 1/1-beta= 10 days and if beta=0.1, then average~= 1.1 meaning, we are giving weightage to the data of the previous day only.